requirements.txt
asgiref==3.10.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dill==0.4.0
Django==5.2.7
django-allauth==65.13.0
django-environ==0.12.0
django-oauth-toolkit==3.1.0
django-storages==1.14.6
djangorestframework==3.16.1
dnspython==2.8.0
drf-spectacular==0.28.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
oauth2_provider==0.0
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pymongo==4.15.3
PyMySQL==1.1.2
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
PyJWT==2.10.1
